+++ composeApp/src/commonMain/kotlin/sharedData/enums/ScreenId.kt
@@ -1,0 +1,5 @@
+package sharedData.enums
+
+enum class ScreenId {
+    Devices, Secrets
+}
